+++ test/resources/get-smiles-response.js
